perf: faster database tab (and better exclude non-tables)
The Database tab in Data Explorer can be quite slow. It's cached, and uses Ajax
on a separate tab to not block the main page load, but it's still not great. By
using PostgreSQL catalogue tables even more directly by leveraging pg_shdepend
we can speed it up.

At busy database times, the previous query took around 45 seconds, and the new
around 1 second. At less busy times the difference is less pronounced, the
previous query taking ~2 seconds and the new around 0.75 seconds.

As a small bonus, this now properly excludes indexes and sequences by using the
'relkind' field when filtering on pg_class.

The change in the tests is because the previous way automatically included
tables granted to the PUBLIC role - the new way doesn't and the user has to
have permissions granted explicitly for them to appear.  However, because in
production we do explicitly grant permissions via catalogue pages, even to
things in the public schema, this makes the new way a touch more similar to
what happens in production.