fix: avoid undoing recent VisualisationLink changes
It's been reported that changing details of VisualisationLinks appear to work,
but then are undone within a few minutes. Although we're not 100% sure, it
could be due to background tasks that query for VisualisationLink instances, do
slow tasks, and then save all the fields of the model instance, and thus
undoing any interim changes to the model instance.

We fix this by using the `update_fields` option of the save method to only save
the specific fields that should be being saved, and thus not undoing any recent
changes persisted in the database.