fix: avoid multiple process_quicksight_dashboard_visualisations
It looks like we're running multiple
process_quicksight_dashboard_visualisations at once - it's set to start every 5
minutes but it might be taking longer than 5 minutes these days.

By wrapping the code that does the work in a lock this should avoid the
problem. This pattern is already in use for other scheduled Celery tasks.