feat: make Arango's nice domain work from tools
The dns-rewrite-proxy rewrites and filters certain DNS requests (it was created
before AWS could do this with their DNS firewall). One of the things it does it
rewrites <host>.<public-domain> to <host>.<private-domain> in tools so that the
same domain can be used from outside of tools as well as in (for example,
GitLab).

However, the arango host under (for example) data.trade.gov.uk resolves to the
internal load balancer, and so we don't need the rewriting, because there is no
arango.<private-domain>.

Also the Docker registry has not existed for several years - so this repurposes
its entry in the rewrite list for Arango.