feat: /pipelines View logs changed to point to data-flow
To simplify the interface between Data Workspace and data-flow, this changes
the "View logs" link in /pipelines from pointing to a page in Data Workspace
that shows a very rough view of the most recent logs for a pipeline to a link
that sends the user to the pipeline page in data-flow.

This is done to make it easier to make each pipeline point to a specific
data-flow while we run 2 data-flow's side by side during the migration off
GOV.UK PaaS, and means we don't have to continue to make it possible to expose
logs via an API in data-flow, which is work because we use a different logging
system in the new data-flow.

Co-authored-by: Peter Woodcock <peter.woodcock@digital.trade.gov.uk>
Co-authored-by: Michal Charemza <michal@charemza.name>