perf: speed up preview permissions checks by avoiding N+1 queries
This changes the source_tables_for_app to be more like the
source_tables_for_user function in that it uses the Django "values" function to
avoid N+1 queries.

This function is used in visualisation previews, and some are being reported as
being very slow - to the point of getting 504 errors. It's not a definite that
this is the main problem, but I think it shouldn't hurt so at worst this should
exclude N+1 as being the problem.

The source_tables_for_user was previously optimised in:
https://github.com/uktrade/data-workspace-frontend/commit/d857b2b92456239d87831298b6fc66d6f3a3b2ee