feat: show pipeline type and platform on pipelines page
This adds the pipeline type (so "Sharepoint" or "SQL") and the platform
("GOV.UK PaaS" or "Data Workspace AWS") to each pipeline in the /pipelines
page.

This is to primarily help the migration off GOV.UK PaaS, so we know what is
done and what's left to do, and is a temporary change. However, the pipeline
type is apparently useful to see, so that might be left longer term.