Fix visualisation losing database permissions
A custom visualisation was reported as losing database permissions
intermittently. Tracked it down to admins of the visualisation causing this
permissions loss when they started a tool.

In more details, when they started a tool, it would cause
`new_private_database_credentials` to run, and as part of that many calls to
`sync_roles`. However, some of some of these were originally designed for team
schemas and their corresponding roles but were being run for visualisation
schemas and roles. This was "bad" because unlike team roles, visualisation
roles are granted SELECT privileges on tables, and the code for team schemas
removed such permissions.

The solution is to just not call `sync_roles` for the visualisation schemas and
roles - it would get run next time the visualisation starts up anyway, and I
suspect this was essentially the behaviour before we introduced `sync_roles` a
few months back.

An alternative would be to run `sync_roles` for the visualisation role, but
pass the full list of tables that a visualisation has access to. However, I
suspect this would be complex and maybe not performant, especially if the user
is admin on a lot of visualisations.