perf: avoid running SET ROLE for contexts that can't make objects
In our PostgreSQL permissions setup, each tool has a temporary database user
that is granted the permanent database role of the person launching the tool.
We also SET ROLE of the temporary user to the permanent role, so that objects
created by the temporary user are then owned by the permanent role.

However, this seems to have performance implications on connecting:

- It increases the "best case" connection time from 0.03s to 0.06s.
- Which doesn't matter too much, but it also increases the frequency of "worst
  case" connection times, when connection takes about 40 seconds.

So this change makes it so we avoid running SET ROLE for the environment where
it's not possible to make objects: QuickSight, Superset and Data Explorer.

This doesn't appear to completely remove "worst case" connection times, but it
is hopefully a small step towards that.