tests: test Your Files team access, including bug when changing email address
This adds a test that the credentials used in Your Files works when a user is
added to a team work to access its team folder, but when the user changes email
address and is added to another team, their credentials don't work for that
team. (This is because the user's role name contains their email address, and
the role name that is stored on their user record doesn't get updated. This is
addressed in a later commit).

This includes updating some of the base64-encoded policies that when deployed
is populated by terraform. These are based on ones from a dev environment but
changing their bucket names (s) to be the ones that are fired up locally and in
tests. It also links LocalStack to the tests in order to more properly test
that policies are changed properly.

As the comment in the tests explains, unfortunately the free LocalStack doesn't
support policy enforcement, so we assert on permissions by a light check on
what the policy text contains. This I think is the best we can reasonably do
for now.