feat: lower risk of undoing user changes when getting credentials
Using a plain "user.save()" runs the risk of undoing recent changes made by
other parts of the application. So we change this to use "update_fields" to
only change the field we are interested in, in this case tools_access_role_arn
that is used when a user access AWS resources directly.

This does not address a reported issue, just something I noticed.