feat: add --prod tag suffix to production visualisations
This adds the "--prod" tag suffix to production Docker-based visualisations.

This is so we can add a lifecycle policy to clean up old previews, but prevent
the production visualisations, that would match "*--prod", from being deleted.

Note that before such a policy is added, we would need to manually add the
"*--prod" tag to all prod images in ECR.