feat: open up CSP for tools to allow VS Code components to communicate
This opens up parts of the CSP on tools to subdomains of vscode-cdn.invalid

This domain can not actually exist, so it doesn't meaningfully increase risk of
data leakage. However, our (modified) VS Code that we run in tools is
configured to request resources (scripts, stylesheets, and fonts) from
subdomain of this domain, and then intercepted by a service worker to retrieve
the contents without communication with the internet.

Why VS Code uses this pattern is not known.

Co-authored-by: Peter Woodcock <peter.woodcock@digital.trade.gov.uk>
Co-authored-by: Michal Charemza <michal@charemza.name>
Co-authored-by: Josef Smith <josef.smith@trade.gov.uk>
Co-authored-by: Sekhar Panja <sekhar.panja@digital.trade.gov.uk>
Co-authored-by: Argyris Galamatis <agalamatis@googlemail.com>