perf: reduce churn on database when creating users
Investigating slow initiation of connections to the database leads to suspected
new_private_database_credentials involvement, and cache effects, so slightly
speculatively looking at reducing what new_private_database_credentials does.

One thing it does is set the default privileges on tables created in shared
(i.e. team) schemas, and in order to do this, grants the current (master) user
the roles to do this. However, an event trigger is in place that automatically
reassigns the ownership of tables creates in such schemas to the corresponding
roles, which by default also gives all privileges on the tables to the owners.

So it shouldn't be necessary to set default privileges, and so also means it
shouldn't be necessary to grant, and then revoke, privileges to the master
user.