feat: separate out roles for tables granted by email domain
This changes how database table privileges are granted for users that get
access by their email domain. Instead of being granted a per-table role, there
is now a per-email-domain role.

This is done to try to limit the number of role memberships users have, which
causes performance problems on connection.

This includes a bump to pg-sync-roles that fixes a bug where it doesn't grant
permissions when it should.