perf: use direct mode for schema create permission
To try to reduce the time to connection to the datasets db, we're reducing the
number of rules the connecting user has.

This change uses direct=True mode on the schema create permission to avoid
using the intermediate role on schemas that users have the ability to create
tables in: private and team schemas.