refactor: fewer data structures
This change simplifies the source_tables_for_user and source_tables_for_app
functions by leveraging Django's F and Values functions to do inline
"normalisation" of results for SourceDatasets and ReferenceDataset, allowing
intermediate data structures (a dictionary for dataset) and list comprehensions
to be removed.

Although the visualisation dataset management area does use the dictionaries,
so they are added back in just for that area.

This is done because we are likely to make the functions source_tables_for_user
and source_tables_for_app more complex in a future change.