feat: open up /pipelines to all users, but have it empty unless superuser
This opens up the /pipeline page to all users, but unless the user is a
superuser, just have it show a message that the user has no access to any
pipelines.

This isn't particularly useful itself - the main reason for this is that this
is a stepping stone in terms of code and UX to a bigger feature where Data
Catalogue Editors, IAMs, and IAOs can trigger pipelines for datasets they
manage.