feat: catalogue editors view own sharepoint pipelines in /pipelines
This adds the ability for data catalogue editors to view the Sharepoint
pipelines for tables in pages they are the catalogue editors for.

This only slightly useful - the main reason for this is that this is a stepping
stone to a bigger feature where they can trigger these pipelines, and that IAMs
and IAOs can trigger their pipelines.