fix: more robust search for managed pipelines defined with quotes
The /pipelines page shows pipelines that the current user can manage, by
essentially finding pipelines that populate tables for source tables in
datasetes for which the current user is an IAO, IAM, or catalogue editor for.
However, /pipelines defines tables in a slightly different way to source
tables, and has to link them fairly "manually".

This change makes the linking more robust, and specifically when the table in
/piplines is specified by identifiers that are quoted.

This still isn't perfect though - there are probably some edge cases where
pipelines would still be missing. The most robust solution is to migrate
/pipelines to accept and store separate schema and table names, but I think
that's a bit of a project, and the chance that anyone will create tables with
names that hit such cases is slim to none, so not worth it right now.