build: fix installing dependencies (#3591)
This

- Downgrades async-timeout in live from 5.0.1 to 4.0.3 (and upgrades in test
  from 4.0.2 to 4.0.3)
- Upgrades gevent from 23.9.1 to 25.5.1

The downgrade of async-timeout - it was odd/"incorrect" that the version was
inconsistent between test and live: they should really be the same version, and
so something had to change. There was some conflicting dependency in test that
forbade the install of the most recent version (I suspect it was aiopg), so a
downgrade in live seemed to be the only option.

The upgrade of gevent is because the older version would not compile with
recent Cython - suspect the same problem as
https://github.com/gevent/gevent/issues/2092. We probably should figure out a
way to pin the version of Cython, but the upgrade of gevent seemed the quickest
way to get things building again.