feat: human readable cron in pipelines page
This makes the schedule descriptions on the pipelines page more human readable
and uses the Cron Descriptor library to address the issue of cron expressions
appearing for custom schedules.

These changes allow users to see when each pipeline is run, especially for
pipelines on custom schedules, which previously were not human readable.